added try catch to service methods

+++ src/main/java/movieticket/services/CinemaService.java
@@ -6,6 +6,7 @@ import java.util.stream.Collectors;
 import movieticket.dtos.CinemaDTO;
 import movieticket.entities.Cinema;
 import movieticket.entities.Movie;
+import movieticket.exceptions.InvalidDataException;
 import movieticket.exceptions.ResourceNotFoundException;
 import movieticket.repositories.CinemaRepository;
 import movieticket.repositories.MovieRepository;
@@ -26,19 +27,33 @@ public class CinemaService {
 	}
 	
 	public void insert(CinemaDTO dto) {
-		Cinema entity = new Cinema();
-		copyDtoToEntity(dto, entity);
-		repository.insert(entity);
+		try {
+			Cinema entity = new Cinema();
+			copyDtoToEntity(dto, entity);
+			repository.insert(entity);
+			System.out.println("Cinema inserido com sucesso: " + dto);
+		} catch (Exception e) {
+			throw new InvalidDataException("Dados inválidos.");
+		}
 	}
 	
 	public void update(Long id, CinemaDTO dto) {
-		Cinema entity = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Id not found"));
-		copyDtoToEntity(dto, entity);
-		repository.update(entity);
+		try {
+			Cinema entity = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Id not found"));
+			copyDtoToEntity(dto, entity);
+			repository.update(entity);
+			System.out.println("Cinema atualizado com sucesso: " + dto);
+		} catch (Exception e) {
+			throw new InvalidDataException("Dados inválidos.");
+		}
 	}
 	
 	public void delete(Long id) {
-		repository.delete(id);
+		CinemaDTO dto = findById(id);
+		if(dto != null) {
+			repository.delete(id);
+			System.out.println("Cinema deletado com sucesso: " + id);
+		}
 	}
 	
 	private void copyDtoToEntity(CinemaDTO dto, Cinema entity) {
