+++ src/main/java/movieticket/services/SeatService.java
@@ -6,6 +6,7 @@ import java.util.stream.Collectors;
 import movieticket.dtos.SeatDTO;
 import movieticket.entities.Seat;
 import movieticket.entities.Ticket;
+import movieticket.exceptions.InvalidDataException;
 import movieticket.exceptions.ResourceNotFoundException;
 import movieticket.repositories.RoomRepository;
 import movieticket.repositories.SeatRepository;
@@ -28,19 +29,33 @@ public class SeatService {
 	}
 	
 	public void insert(SeatDTO dto) {
-		Seat entity = new Seat();
-		copyDtoToEntity(dto, entity);
-		repository.insert(entity);
+		try {
+			Seat entity = new Seat();
+			copyDtoToEntity(dto, entity);
+			repository.insert(entity);
+			System.out.println("Assento inserido com sucesso: " + dto);
+		} catch(Exception e) {
+			throw new InvalidDataException("Dados inválidos.");
+		}
 	}
 	
 	public void update(Long id, SeatDTO dto) {
-		Seat entity = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Id not found"));
-		copyDtoToEntity(dto, entity);
-		repository.update(entity);
+		try {
+			Seat entity = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Id not found"));
+			copyDtoToEntity(dto, entity);
+			repository.update(entity);
+			System.out.println("Assento atualizado com sucesso: " + dto);
+		} catch(Exception e) {
+			throw new InvalidDataException("Dados inválidos.");
+		}
 	}
 	
 	public void delete(Long id) {
-		repository.delete(id);
+		SeatDTO dto = findById(id);
+		if(dto != null) {
+			repository.delete(id);
+			System.out.println("Cinema deletado com sucesso: " + id);
+		}
 	}
 	
 	private void copyDtoToEntity(SeatDTO dto, Seat entity) {
